[FS-257] - feat: hook to handle form created
